Correct italic n, superscript and subscript
In the original MIM file, the character combination "in" results in the symbol ∈ and not the italicized 𝑛, because the two combinations clashed. In the first change, the character combination for ∈ was altered to "𝐈𝐧" (with an uppercase 𝐈)

Secondly, unicode characters for superscript and subscript were incorporated into the MIM file, using "^2" for ² for instance. This allows the user to (1) write chemical equations:
2H₂O → 2H₂ + O₂ using "_2", for instance
and (2) write complex equations:
xₙ² - x₀ⁿ = x₀⁽ⁿ⁺¹⁾
and also (3) compose n-ary operations and integrations:
Σᵢ aᵢ² = A